fix: formatdate helper function
